Refactor: Refactored instance status struct and methods
