+++ internal/inventory/types.go
@@ -1,24 +1,8 @@
 package inventory
 
-import (
-	"strings"
-)
-
-// InstanceStatus defines the status of the instance
-type InstanceStatus string
-
 const (
 	// ClusterTagKey string to identify to which cluster is the instance associated
 	ClusterTagKey string = "kubernetes.io/cluster/"
-)
-
-const (
-	// Running Instance status
-	Running InstanceStatus = "Running"
-	// Stopped Instance status
-	Stopped InstanceStatus = "Stopped"
-	// Terminated Instance status
-	Terminated InstanceStatus = "Terminated"
 )
 
 const (
@@ -30,19 +14,3 @@ const (
 	ClusterResourceType  = "cluster"
 	InstanceResourceType = "instance"
 )
-
-// AsInstanceStatus converts the incoming argument into a InstanceStatus type
-func AsInstanceStatus(status string) InstanceStatus {
-	switch strings.ToLower(status) {
-	case "running":
-		return Running
-	case "stop":
-		return Stopped
-	case "stopped":
-		return Stopped
-	case "terminated":
-		return Terminated
-	default:
-		return Running
-	}
-}
